refactors mapTableColumnsToEuiColumns
